Parallel v1.1.0 (#20)
* update to V1.1

* update to V1

* update to V1.1

* update to V1.1

* update to V1.1

* add test environment to path

* Update to V1.1 - Parallelization

* add sample dependent error reporting

* add execution indepdent of submission

* updated bash clean script to current folder structure

* updated bash clean script to current folder structure

* update to V1.1

* fix sample order bug

* fix O2 run

* add displaying sample name

* fix path bug

* fix for loop bug

* fix individual sample

* update to single image from directory

* fix UNET bug

* add CODEX code

* fix indent

* fix Histocat bug

* remove test lines

* fix segmentation bug

* fix segmentation bug

* update error log

* update error log

* fix HistoCat comma

* Update to V1.1

* Update to V1.1: Parallelization

* Update to V1.1: Parallelization
