+++ bin/cycif_pipeline_activate.sh
@@ -1,4 +1,4 @@
 module load conda2/4.2.13
-source activate /n/groups/lsp/cycif/cycif_pipeline/
-python /n/groups/lsp/cycif/CyCif_Manager/O2/CyCif_Pipeline_O2_v1.py $1
+source activate /n/groups/lsp/cycif/cycif_pipeline_testing_space/mcmicro/environments/cycif_pipeline/
+python /n/groups/lsp/cycif/cycif_pipeline_testing_space/mcmicro/O2/CyCif_Pipeline_O2_v1.1.py $1
 conda deactivate
