+++ install_example_dataset.sh
@@ -1,5 +1,5 @@
 #Download example data for testing pipeline
 
-wget --output-document=example_data.tar.gz https://www.dropbox.com/s/x0m1hc5m0we8bfv/example_data.tar.gz?dl=0
+wget --output-document=example_data.tar.gz https://www.dropbox.com/s/e3c3eqe3azle62f/example_data.tar.gz?dl=0
 tar -zxf example_data.tar.gz
 rm example_data.tar.gz
